Sanitize common metrics tag values, allow dashes
Dashes are already used in names in other locations, so it's fairly implied that they're safe.

Dots I'm excluding intentionally, as it's used as a conceptual-group separator, e.g. as in canary/metrics.go.
Since these tag APIs seem more focused on single-"thing" values, separators don't seem like they should be allowed.

Also, as the existence of a sanitizer implies that there are *un*safe values, we should probably always sanitize.
So now everything's sanitized.  I've excluded the kafka one simply because the current builder + sanitizer are guaranteed to not have any effect.
